remove print statement of available functions at startup
This new print statement has resulted in users confused about the new print out in the console. It also does not seem that informative, since the information is the same on each running of the assistant (unless of course a new compatible library is installed). 

Since this is useful information on the developer side, then consider allowing it to be called from a separate method. Maybe this was just for debugging purposes at the time.